+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Improvement: Increase diagnostic response payload size to accommodate
   multi-frame diagnostic responses.
+* Improvement: Added command to request device platform
 
 ## v0.6.0